# internal/models/user.go
package models

import "time"

type User struct {
	ID             uint   `gorm:"primaryKey"`
	PassportNumber string `gorm:"unique; not null"`
	Surname        string `gorm:"not null"`
	Name           string `gorm:"not null"`
	Patronymic     string
	Address        string `gorm:"not null"`
	CreatedAt      time.Time
	UpdatedAt      time.Time
}